support heterogeneous machines, eliminating FIXME
The detection algorithm is: determine the number of processing elements
available to us, then iterate over each one and detect its details.
Check the details against those we've already seen; if they're
different, that's a new processor type. The full skeleton is now
present, with several FIXMEs: we want to use cpuset functionality, not
this ghetto sysconf(). we need the actual details detection. etc.