share event queues, fixing bug 57
this is pretty much the second major objective (although an incomplete
implementation), hurrah. we can't have an event queue per thread in the
limit (that's 1k file descriptors on a 16-way SMT, 64-core machine), and
we don't necessarily want one even when we can afford it. this is a lot
of proof of concept on linux regarding multiple threads and epoll -- the
internal implementation just wraps things with a futex, so all those
semantics are inherited. ahhh, this feels really good.