fix up uninitialized access
detected this via the 'hardtest' target's valgrind run. we were using |=
on ee.events without ever having specified an initial value thereof. it's
unlikely that this could have broken anything (this code was specific to
signal setup, which isn't exactly going to generate an EPOLLOUT event),
but poor form. we're back to a clean valgrind (when libnuma's disabled,
anyway)